+++ UsefulBombs/Changelog.md
@@ -3,4 +3,12 @@
 - Fixed KeyNotFoundException.
 
 ## 1.0.2
-- Added 
+- Added  'Cancel Enemy Invincibility'
+
+## 1.0.3
+- Added 'Instant Kill Mummies'.
+- Fixed Illegal items spawning.
+
+## 1.0.4
+- Fixed causing infinite explosion loop outside of mines.
+- Fixed sometimes odd sprites appearing after exploding.